# myits-broadcast-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for myITS Broadcast Fullstack Platform.",
        "title": "myITS Broadcast API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/broadcasts": {
            "get": {
                "description": "Fetch all broadcast records from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Broadcasts"
                ],
                "summary": "Get all active broadcasts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new broadcast entry in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Broadcasts"
                ],
                "summary": "Create a new broadcast",
                "parameters": [
                    {
                        "description": "Broadcast Data",
                        "name": "broadcast",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Broadcast"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/broadcasts/{id}": {
            "get": {
                "description": "Fetch a single broadcast record by its primary key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Broadcasts"
                ],
                "summary": "Get broadcast by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Broadcast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a broadcast record from database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Broadcasts"
                ],
                "summary": "Delete broadcast by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Broadcast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/broadcasts/{id}/click": {
            "post": {
                "description": "Increment the CTA click count for a specific broadcast",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Broadcasts"
                ],
                "summary": "Increment click count",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Broadcast ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Broadcast": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "author_email": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "clicks": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "cta_type": {
                    "type": "string"
                },
                "cta_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expiry_date": {
                    "type": "string"
                },
                "expiry_days": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "description": "\u003c--- Field Baru",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}